# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.FullChat:
    properties:
      avatarPath:
        type: string
      id:
        type: integer
      isGroup:
        type: boolean
      lastMessage:
        $ref: '#/definitions/domain.Message'
      lastMessageAuthor:
        $ref: '#/definitions/domain.ShortProfile'
      lastReadMessageID:
        type: integer
      name:
        type: string
      unReadCounts:
        type: integer
    type: object
  domain.Message:
    properties:
      authorID:
        type: integer
      chatID:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      text:
        type: string
    type: object
  domain.MessagesWithAuthors:
    properties:
      authors:
        additionalProperties:
          $ref: '#/definitions/domain.ShortProfile'
        type: object
      messages:
        items:
          $ref: '#/definitions/domain.Message'
        type: array
    type: object
...
  domain.ShortProfile:
    properties:
      avatarPath:
        type: string
      dob:
        type: string
      fullName:
        type: string
      userID:
        type: integer
    type: object
...
  handler.ChatIDResponse:
    properties:
      chatID:
        type: integer
    type: object
...
  handler.JSONResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  handler.MessageIDResponse:
    properties:
      messageID:
        type: integer
    type: object
  handler.UpdateLastReadRequest:
    properties:
      lastReadMessageID:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a VK API.
  termsOfService: http://swagger.io/terms/
  title: VK API
  version: "1.0"
paths:
...
  /chats:
    get:
      consumes:
      - application/json
      description: Возвращает постраничный список чатов для текущего (аутентифицированного)
        пользователя.
      parameters:
      - default: 20
        description: Количество чатов для возврата
        format: int32
        in: query
        name: limit
        type: integer
      - default: 0
        description: Номер страницы для пагинации
        format: int32
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список чатов
          schema:
            items:
              $ref: '#/definitions/domain.FullChat'
            type: array
        "400":
          description: Некорректные параметры запроса
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Получить список чатов пользователя
      tags:
      - chats
  /chats/{chatID}/message:
    post:
      consumes:
      - application/json
      description: Создает новое сообщение в указанном чате
      parameters:
      - description: ID чата
        format: int32
        in: path
        name: chatID
        required: true
        type: integer
      - description: Тело сообщения
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/domain.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Создать сообщение
      tags:
      - messages
  /chats/{id}/last-read:
    put:
      consumes:
      - application/json
      description: Обновляет значение lastReadMessageID для текущего (аутентифицированного)
        пользователя в указанном чате.
      parameters:
      - description: ID чата
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - description: Новый ID последнего прочитанного сообщения
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateLastReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Информация об успешном обновлении или отсутствии изменений
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Некорректные параметры запроса
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Обновить последнее прочитанное сообщение
      tags:
      - chats
  /chats/{id}/messages:
    get:
      description: Возвращает сообщения из чата с пагинацией и краткую информацию
        об авторах
      parameters:
      - description: ID чата
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - default: 20
        description: Лимит количества сообщений
        format: int32
        in: query
        name: limit
        type: integer
      - default: 0
        description: страница для пагинации
        format: int32
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessagesWithAuthors'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Получить сообщения чата с авторами
      tags:
      - messages
  /chats/user/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает ID чата между текущим пользователем и указанным userID,
        создавая чат при отсутствии
      parameters:
      - description: ID пользователя
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChatIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Получить или создать чат с пользователем
      tags:
      - chats
...
swagger: "2.0"